pyproject.toml
[project]
name = "CS2_Arb"                       
version = "0.1.0"
description = "CS2 skin arbitrage tool"
requires-python = ">=3.11"
authors = [{ name = "Jacob Hwang" }]
license = { text = "Proprietary" }

# Python packages you actually need
dependencies = [
  "httpx==0.27.2",          # HTTP client for CSFloat API
  "python-dotenv==1.0.1",   # load .env variables
  "tabulate==0.9.0",        # pretty tables in console
  "pandas==2.2.3"           # optional for CSV/logging
]

# These sections configure your formatting & linting tools
[tool.black]
line-length = 100
target-version = ["py311"]
skip-string-normalization = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I"]
fix = true

# Optional: define default build system (standard)
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"